lab 7: fixed lab7

+++ apps/bookmodule/models.py
@@ -4,4 +4,4 @@ class Book(models.Model):
     title = models.CharField(max_length=50)
     author = models.CharField(max_length=50)
     price = models.FloatField(default=0.0)
-    edition = models.IntegerField(default=1)
+    edition = models.SmallIntegerField(default=1)
